Update patch set 4
Patch Set 4: (6 inline comments)

Took care of most of the problems. I'll hold off on writing the docs until we decide if Martin's idea is a good one :)

Martin: that actually sounds like a great idea to me. Quick list of possible commands:
`branch create`
`branch delete`
`branch change-head`

It'd be nice if the GUI could take care of the change-head scenario as well, so we've got feature parity here.

Patch-set: 4